#41-`TripService`레이어 분리
