+++ src/main/java/com/fivefeeling/memory/domain/media/repository/MediaFileRepository.java
@@ -10,8 +10,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface MediaFileRepository extends JpaRepository<MediaFile, Long> {
 
-  List<MediaFile> findByTripTripId(Long tripId);
-
   List<MediaFile> findByPinPointPinPointId(Long pinPointId);
 
   // tripId와 pinPointId에 해당하는 MediaFile 조회
