+++ src/main/java/com/fivefeeling/memory/domain/trip/service/TripQueryService.java
@@ -13,10 +13,7 @@ import com.fivefeeling.memory.domain.trip.model.PointImageDTO;
 import com.fivefeeling.memory.domain.trip.model.Trip;
 import com.fivefeeling.memory.domain.trip.model.TripDetailsDTO;
 import com.fivefeeling.memory.domain.trip.model.TripInfoDTO;
-import com.fivefeeling.memory.domain.trip.model.TripRequestDTO;
-import com.fivefeeling.memory.domain.trip.model.TripResponseDTO;
 import com.fivefeeling.memory.domain.trip.model.TripSummaryDTO;
-import com.fivefeeling.memory.domain.trip.model.TripUdateRequestDTO;
 import com.fivefeeling.memory.domain.trip.repository.TripRepository;
 import com.fivefeeling.memory.domain.user.model.User;
 import com.fivefeeling.memory.domain.user.model.UserTripsDTO;
@@ -33,7 +30,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class TripService {
+public class TripQueryService {
 
   private final TripRepository tripRepository;
   private final PinPointRepository pinPointRepository;
@@ -42,9 +39,7 @@ public class TripService {
 
   private static final String TRIP_NOT_FOUND = "해당 여행이 존재하지 않습니다.";
   private static final String MEDIA_FILE_NOT_FOUND = "해당 핀포인트의 미디어 파일이 존재하지 않습니다.";
-  // 날짜 포맷 형식 지정
   private static final SimpleDateFormat DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd");
-
 
   public List<TripSummaryDTO> getTripsByUserId(Long userId) {
     return tripRepository.findByUserUserId(userId).stream()
@@ -55,9 +50,13 @@ public class TripService {
         .collect(Collectors.toList());
   }
 
+
   public List<PinPointSummaryDTO> getPinPointsByUserId(Long userId) {
     return tripRepository.findByUserUserId(userId).stream()
-        .map(trip -> pinPointRepository.findByTripTripId(trip.getTripId()).stream().findFirst().orElse(null))
+        .map(trip -> pinPointRepository.findByTripTripId(trip.getTripId())
+            .stream()
+            .findFirst()
+            .orElse(null))
         .filter(pinPoint -> pinPoint != null)
         .map(pinPoint -> new PinPointSummaryDTO(
             pinPoint.getTrip().getTripId(),
@@ -68,29 +67,6 @@ public class TripService {
         .collect(Collectors.toList());
   }
 
-  public TripResponseDTO createTrip(String userEmail, TripRequestDTO tripRequestDTO) {
-    User user = userRepository.findByUserEmail(userEmail)
-        .orElseThrow(() -> new IllegalArgumentException("해당 사용자가 존재하지 않습니다."));
-    Trip trip = Trip.builder()
-        .user(user)
-        .tripTitle(tripRequestDTO.tripTitle())
-        .country(tripRequestDTO.country())
-        .startDate(tripRequestDTO.startDate())
-        .endDate(tripRequestDTO.endDate())
-        .build();
-
-    trip.setHashtagsFromList(tripRequestDTO.hashtags());
-    Trip savedTrip = tripRepository.save(trip);
-
-    return new TripResponseDTO(
-        savedTrip.getTripId(),
-        savedTrip.getTripTitle(),
-        savedTrip.getCountry(),
-        savedTrip.getStartDate(),
-        savedTrip.getEndDate(),
-        savedTrip.getHashtagsAsList()
-    );
-  }
 
   public UserTripsDTO getUserTripInfo(String userEmail) {
     User user = userRepository.findByUserEmail(userEmail)
@@ -109,44 +85,6 @@ public class TripService {
     return new UserTripsDTO(user.getUserNickName(), trips);
   }
 
-  // 사용자 여행 정보 수정
-  @Transactional
-  public TripResponseDTO updateTrip(String userEmail, Long tripId, TripUdateRequestDTO tripUdateRequestDTO) {
-    User user = userRepository.findByUserEmail(userEmail)
-        .orElseThrow(() -> new IllegalArgumentException("해당 사용자가 존재하지 않습니다."));
-
-    Trip trip = tripRepository.findById(tripId)
-        .orElseThrow(() -> new IllegalArgumentException("해당 여행이 존재하지 않습니다."));
-
-    trip.setTripTitle(tripUdateRequestDTO.tripTitle());
-    trip.setCountry(tripUdateRequestDTO.country());
-    trip.setStartDate(tripUdateRequestDTO.startDate());
-    trip.setEndDate(tripUdateRequestDTO.endDate());
-    trip.setHashtagsFromList(tripUdateRequestDTO.hashtags());
-
-    Trip updatedTrip = tripRepository.save(trip);
-
-    return new TripResponseDTO(
-        updatedTrip.getTripId(),
-        updatedTrip.getTripTitle(),
-        updatedTrip.getCountry(),
-        updatedTrip.getStartDate(),
-        updatedTrip.getEndDate(),
-        updatedTrip.getHashtagsAsList()
-    );
-  }
-
-  // 사용자 여행 정보 삭제
-  @Transactional
-  public void deleteTrip(String userEmail, Long tripId) {
-    User user = userRepository.findByUserEmail(userEmail)
-        .orElseThrow(() -> new IllegalArgumentException("해당 사용자가 존재하지 않습니다."));
-
-    Trip trip = tripRepository.findById(tripId)
-        .orElseThrow(() -> new IllegalArgumentException("해당 여행이 존재하지 않습니다."));
-
-    tripRepository.delete(trip);
-  }
 
   public TripDetailsDTO getTripInfoById(Long tripId) {
     Trip trip = tripRepository.findById(tripId)
@@ -185,9 +123,11 @@ public class TripService {
     );
   }
 
+
   private String formatDate(Date date) {
     return date != null ? DATE_FORMAT.format(date) : null;
   }
+
 
   // Pinpoint 슬라이드 쇼 조회
   @Transactional(readOnly = true)
